requirements.txt
# Основные зависимости Django
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
django-cors-headers==4.2.0
django-filter==23.3

# База данных и ORM
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-dotenv==1.0.0

# Медиафайлы и изображения
Pillow==10.1.0

# Интернационализация и локализация
django-modeltranslation==0.24.0
django-parler==2.3
django-parler-rest==2.2
Babel==2.13.1

# Безопасность и аутентификация
sentry-sdk==1.38.0
django-axes==6.1.1
django-defender==0.9.6

# Кэширование и очереди
django-redis==5.3.0
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Уведомления и коммуникации
firebase-admin==6.2.0
django-push-notifications==3.0.0

# Платежи
stripe==5.0.0

# Геолокация и карты
geopy==2.4.0
django-geoposition==0.3.0

# Документация API
swagger-spec-validator==3.0.3

# Мониторинг и логирование
django-log-request-id==2.0.0

# Тестирование
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2

# Утилиты
python-dateutil==2.8.2

# Деплой
gunicorn==21.2.0
whitenoise==6.6.0

# Разработка и отладка
ipython==8.17.2
django-extensions==3.2.3
werkzeug==3.0.1
setuptools==69.0.2
